Correctly include `poem` class on poem layout
BackstopJS is already paying off big time here, because without it I
might not have caught this little subtle issue.

While the `poem.html` layout used `poetry` as its container class, the
CSS referenced a `poem` class (which fits with narrative and drama using
their same file names as the container class).

However, because `div.poem` did not exist, the margin set on the bottom
of poems was never applied. That is now fixed, which is why the
reference images are updated with this commit.

(That said, the margins are collapsing in a way the original author of
the CSS likely did not intend, but that's a fix for another day.)

To preserve backward-compatibility, I have added the `poem` class to the
existing `poetry` class in the poem template. I have also updated the
CSS to account for the use of either `poem` or `poetry` as the container
class names.