Simplify and improve the 404 page
This removes the old `404.html` file (which Jekyll overwrote with the
contents of `404.md` anyway), and properly uses the `page` layout that
the old `404.md` was shoehorning into the content area.

With the `page` layout in effect, the sidebar has been modified not to
list the 404 page in the site contents.